requirements.txt
boto3==1.10.2
botocore==1.13.2
certifi==2019.9.11
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
gunicorn==20.0.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
mysql-connector-python==8.0.18
protobuf==3.10.0
python-dateutil==2.8.0
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
